+++ eve/tests/methods/delete.py
@@ -559,26 +559,24 @@ class TestSoftDelete(TestDelete):
             # show_deleted == True is passed or if the deleted field is part of
             # the lookup
             req.show_deleted = False
-            docs = self.app.data.find(self.known_resource, req, None)
-            undeleted_count = docs.count()
+            self.app.data.find(self.known_resource, req, None)
+            undeleted_count = self.app.data.last_documents_count
 
             req.show_deleted = True
-            docs = self.app.data.find(self.known_resource, req, None)
-            with_deleted_count = docs.count()
-            self.assertEqual(undeleted_count, with_deleted_count - 1)
+            self.app.data.find(self.known_resource, req, None)
+            self.assertEqual(undeleted_count, self.app.data.last_documents_count - 1)
 
             req.show_deleted = False
-            docs = self.app.data.find(
-                self.known_resource, req, {self.deleted_field: True}
-            )
-            deleted_count = docs.count()
+            self.app.data.find(self.known_resource, req, {self.deleted_field: True})
+            deleted_count = self.app.data.last_documents_count
             self.assertEqual(deleted_count, 1)
 
             # find_list_of_ids will return deleted documents if given their id
-            docs = self.app.data.find_list_of_ids(
+            self.app.data.find_list_of_ids(
                 self.known_resource, [ObjectId(self.item_id)]
             )
-            self.assertEqual(docs.count(), 1)
+
+            self.assertEqual(self.app.data.last_documents_count, 1)
 
     def test_softdelete_db_fields(self):
         """Documents created when soft delete is enabled should include and